refactor: /src/components 내부 디렉토리 생성 및 import 경로 수정
/src/components 내부에 /common, /review 디렉토리 생성하여 컴포넌트를 분류하고 import 경로를 수정함
- /common : 페이지에 관계없이 공통으로 사용할 수 있는 컴포넌트
- /review : 리뷰 페이지에서만 사용할 컴포넌트
