+++ src/views/components/common/SpeechBubbleStudent.tsx
