+++ src/views/components/common/SuggestionCard.tsx
