+++ src/views/components/common/WordCard.tsx
