+++ src/views/components/review/ReviewBody.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Grid } from "@material-ui/core";
-import Container from "./Container";
+import Container from "@/common/Container";
 import * as options from "Const/containerOptions";
 
 const ReviewBody: React.FC = () => (
